+++ src/App.tsx
@@ -9,7 +9,7 @@ import { RouterProvider, createBrowserRouter } from 'react-router-dom';
 import { routes } from './pages/routes';
 import ThemeCustomization from './themes/themeCustomization';
 import { UserContext } from './components/contexts/UserContext';
-import { SpinUpProvider } from './components/contexts/SpinUpContext';
+import { SpinUpProvider } from './components/contexts/SpinUpProvider';
 import { ClientPrincipal, User } from './types/interfaces';
 import usePersistentState from './hooks/usePersistentState';
 import useAuthorization from './hooks/useAuthorization';
