+++ src/components/contexts/spinUpContext.ts
@@ -5,12 +5,6 @@
  *
  */
 import { createContext } from 'react';
-
-export interface SpinUpContextType {
-  showDialog: boolean;
-  retryCount: number;
-  setShowDialog: (show: boolean) => void;
-  setRetryCount: (count: number) => void;
-}
+import { SpinUpContextType } from '../../types/interfaces';
 
 export const SpinUpContext = createContext<SpinUpContextType | undefined>(undefined);
